+++ src/css/form/list.ts
@@ -1,0 +1,11 @@
+import { css, CSSResult } from "lit";
+
+export const listStyle: CSSResult = css`
+
+  ul {
+    list-style-type:          none;
+    padding:                  0;
+    margin:                   0;
+  }
+
+`;