+++ src/custom-elements/drawer/css/link.ts
@@ -1,0 +1,15 @@
+
+import { css, CSSResult } from 'lit';
+
+export const linkStyle: CSSResult = css`
+
+  a {
+    text-transform:           uppercase;
+    font-size:                .6em;
+    color:                    rgb(105,105,105);
+    font-weight:              800;
+    text-decoration:          none;
+  }
+
+`;
+