+++ src/custom-elements/drawer/css/paragraph.ts
@@ -1,0 +1,12 @@
+import { css, CSSResult } from "lit";
+
+export const lineStyle: CSSResult = css`
+
+  p {
+    text-align: center;
+  }
+  
+  .setLog > p   { margin: 12px; box-sizing: border-box; }
+  .divide       { border: 2px solid #606060; margin: 6px 0 0 0; border-radius: 50%; }
+
+`;